Seperation of power(xd)
